Cleanup sui system state file
