+++ crates/sui-types/src/collection_types.rs
@@ -4,6 +4,8 @@
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
+use crate::base_types::{ObjectID, SuiAddress};
+
 /// Rust version of the Move sui::vec_map::VecMap type
 #[derive(Debug, Serialize, Deserialize, Clone, Eq, PartialEq, JsonSchema)]
 pub struct VecMap<K, V> {
@@ -21,4 +23,69 @@ pub struct Entry<K, V> {
 #[derive(Debug, Serialize, Deserialize, Clone, Eq, PartialEq, JsonSchema)]
 pub struct VecSet<T> {
     contents: Vec<T>,
+}
+
+/// Rust version of the Move std::option::Option type.
+#[derive(Debug, Serialize, Deserialize, Clone, Eq, PartialEq, JsonSchema)]
+pub struct MoveOption<T> {
+    pub vec: Vec<T>,
+}
+
+impl<T> MoveOption<T> {
+    pub fn empty() -> Self {
+        Self { vec: vec![] }
+    }
+}
+
+/// Rust version of the Move sui::table::Table type.
+#[derive(Debug, Serialize, Deserialize, Clone, Eq, PartialEq, JsonSchema)]
+pub struct TableVec {
+    pub contents: Table,
+}
+
+impl Default for TableVec {
+    fn default() -> Self {
+        TableVec {
+            contents: Table {
+                id: ObjectID::from(SuiAddress::ZERO),
+                size: 0,
+            },
+        }
+    }
+}
+
+/// Rust version of the Move sui::table::Table type.
+#[derive(Debug, Serialize, Deserialize, Clone, Eq, PartialEq, JsonSchema)]
+pub struct Table {
+    pub id: ObjectID,
+    pub size: u64,
+}
+
+impl Default for Table {
+    fn default() -> Self {
+        Table {
+            id: ObjectID::from(SuiAddress::ZERO),
+            size: 0,
+        }
+    }
+}
+
+/// Rust version of the Move sui::linked_table::LinkedTable type.
+#[derive(Debug, Serialize, Deserialize, Clone, Eq, PartialEq, JsonSchema)]
+pub struct LinkedTable<K> {
+    pub id: ObjectID,
+    pub size: u64,
+    pub head: MoveOption<K>,
+    pub tail: MoveOption<K>,
+}
+
+impl<K> Default for LinkedTable<K> {
+    fn default() -> Self {
+        LinkedTable {
+            id: ObjectID::from(SuiAddress::ZERO),
+            size: 0,
+            head: MoveOption { vec: vec![] },
+            tail: MoveOption { vec: vec![] },
+        }
+    }
 }
